requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Machine Learning & AI
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# Vector Database
pinecone-client==2.2.4

# OCR
pytesseract==0.3.10
Pillow==10.0.0

# Image Processing
opencv-python==4.8.0.76

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.11.2
requests==2.31.0

# Data Processing
scipy==1.11.1

# Environment Management
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Development
jupyter==1.0.0
matplotlib==3.7.2
seaborn==0.12.2